Split integration tests into multiple files to avoid OOMs.
